+++ frontend/public/locales/ko/settings.json
@@ -13,8 +13,7 @@
         "appearance": "테마",
         "reactions": "리액션",
         "blocks": "차단",
-        "info": "정보",
-        "sign_out": "로그아웃"
+        "info": "정보"
     },
     "account": {
         "display_name": "이름",
@@ -53,6 +52,12 @@
         "delete_task_after_alert": {
             "name": "작업 삭제 시 경고",
             "description": "작업을 삭제할 때 되묻는 경고창을 띄웁니다."
+        },
+        "sign_out": {
+            "name": "여기서 로그아웃",
+            "values": {
+                "button_sign_out": "로그아웃"
+            }
         }
     },
     "security": {
