+++ frontend/src/pages/settings/General.jsx
@@ -1,15 +1,29 @@
-import { useMemo } from "react"
+import { startTransition, useMemo, useState } from "react"
 
+import styled from "styled-components"
+
+import Button from "@components/common/Button"
+import ConfirmationSignOut from "@components/settings/ConfirmationSignOut"
 import Section, { Description, Name, Value } from "@components/settings/Section"
 import Select from "@components/settings/Select"
 import SettingSwitch from "@components/settings/SettingSwitch"
 
+import FeatherIcon from "feather-icons-react"
 import { useTranslation } from "react-i18next"
 
 const General = () => {
     const { t } = useTranslation("settings", { keyPrefix: "general" })
 
     const startpageChoices = useMemo(() => makeStartpageChoices(t), [t])
+
+    const [isSignOutConfirmationOpen, setSignOutConfirmationOpen] =
+        useState(false)
+
+    const openSignOutConfirmation = () => {
+        startTransition(() => {
+            setSignOutConfirmationOpen(true)
+        })
+    }
 
     return (
         <>
@@ -30,6 +44,21 @@ const General = () => {
                     <SettingSwitch name="delete_task_after_alert" />
                 </Value>
             </Section>
+
+            <Section>
+                <Name>{t("sign_out.name")}</Name>
+                <Value>
+                    <Button onClick={openSignOutConfirmation}>
+                        <SignOutIcon icon="log-out" />
+                        {t("sign_out.values.button_sign_out")}
+                    </Button>
+                </Value>
+                {isSignOutConfirmationOpen && (
+                    <ConfirmationSignOut
+                        onClose={() => setSignOutConfirmationOpen(false)}
+                    />
+                )}
+            </Section>
         </>
     )
 }
@@ -44,5 +73,10 @@ const makeStartpageChoices = (t) => [
         value: "today",
     },
 ]
+
+const SignOutIcon = styled(FeatherIcon)`
+    stroke-width: 3px;
+    top: 0;
+`
 
 export default General
